+++ docs/conf.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # Configuration file for the Sphinx documentation builder.
 #
 # This file does only contain a selection of the most common options. For a
@@ -12,24 +10,25 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 
-import os
+import pathlib
 import sys
 
 import tomli
 
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, str(pathlib.Path('..').resolve()))
 
 
 # -- Project information -----------------------------------------------------
 
+
 def _get_project_meta():
-    with open('../pyproject.toml', mode='rb') as pyproject:
+    with pathlib.Path('../pyproject.toml').open(mode='rb') as pyproject:
         return tomli.load(pyproject)['tool']['poetry']
 
 
 pkg_meta = _get_project_meta()
 project = str(pkg_meta['name'])
-copyright = '2018, wemake.services'  # noqa: WPS125
+copyright = '2018, wemake.services'  # noqa: A001
 author = 'wemake.services'
 
 # The short X.Y version
@@ -39,8 +38,6 @@ release = version
 
 
 # -- General configuration ---------------------------------------------------
-
-needs_sphinx = '5.3'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -53,13 +50,10 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.autosummary',
     'sphinx.ext.napoleon',
-
     # Used to include .md files:
-    'm2r2',
-
+    'myst_parser',
     # Used to insert typehints into the final docs:
     'sphinx_autodoc_typehints',
-
     # Used to document `click` commands:
     'sphinx_click.ext',
 ]
@@ -111,42 +105,9 @@ autodoc_default_options = {
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'alabaster'
-
-# Theme options are theme-specific and customize the look and feel of a theme
-# further.  For a list of options available for each theme, see the
-# documentation.
-html_theme_options = {
-    'sidebar_collapse': False,
-    'show_powered_by': False,
-}
+html_theme = 'furo'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
-
-# Custom sidebar templates, must be a dictionary that maps document names
-# to template names.
-#
-# This is required for the alabaster theme
-# refs: http://alabaster.readthedocs.io/en/latest/installation.html#sidebars
-html_sidebars = {
-    '**': [
-        'about.html',
-        'navigation.html',
-        'moreinfo.html',
-        'github.html',
-        'searchbox.html',
-    ],
-}
-
-
-# -- Extension configuration -------------------------------------------------
-
-napoleon_numpy_docstring = False
-
-# -- Options for todo extension ----------------------------------------------
-
-# If true, `todo` and `todoList` produce output, else they produce nothing.
-todo_include_todos = True