+++ dotenv_linter/constants.py
@@ -1,6 +1,5 @@
 """This module contains list of black-listed ``environment`` variables."""
 
-
 from typing import Final
 
 #: List of variable we forbid to use.