+++ dotenv_linter/grammar/parsetab.py
@@ -27,13 +27,13 @@ for _k, _v in _lr_goto_items.items():
 del _lr_goto_items
 _lr_productions = [
   ("S' -> body","S'",1,None,None,None),
-  ('body -> <empty>','body',0,'p_body','parser.py',64),
-  ('body -> body line','body',2,'p_body','parser.py',65),
-  ('line -> assign','line',1,'p_line','parser.py',71),
-  ('line -> name','line',1,'p_line','parser.py',72),
-  ('line -> comment','line',1,'p_line','parser.py',73),
-  ('assign -> NAME EQUAL','assign',2,'p_assign','parser.py',79),
-  ('assign -> NAME EQUAL VALUE','assign',3,'p_assign','parser.py',80),
-  ('name -> NAME','name',1,'p_name','parser.py',85),
-  ('comment -> COMMENT','comment',1,'p_comment','parser.py',89),
+  ('body -> <empty>','body',0,'p_body','parser.py',69),
+  ('body -> body line','body',2,'p_body','parser.py',70),
+  ('line -> assign','line',1,'p_line','parser.py',78),
+  ('line -> name','line',1,'p_line','parser.py',79),
+  ('line -> comment','line',1,'p_line','parser.py',80),
+  ('assign -> NAME EQUAL','assign',2,'p_assign','parser.py',86),
+  ('assign -> NAME EQUAL VALUE','assign',3,'p_assign','parser.py',87),
+  ('name -> NAME','name',1,'p_name','parser.py',97),
+  ('comment -> COMMENT','comment',1,'p_comment','parser.py',101),
 ]